+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Add `Middleware\RouteInterface` and `RequestHandler\RouteInterface`
+
 ### Changed
 - All classes are `final`.
 
@@ -49,7 +52,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 First release
 
-[Unreleased]: https://github.com/ajgarlag/psr15-router/compare/0.4.0...master
+[Unreleased]: https://github.com/ajgarlag/psr15-router/compare/0.5.0...master
+[0.5.0]: https://github.com/ajgarlag/psr15-router/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/ajgarlag/psr15-router/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/ajgarlag/psr15-router/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/ajgarlag/psr15-router/compare/0.1.1...0.2.0